Add postprocessor for joining sections
This change allows OmegaDoc to correctly parse section definitions in
OmegaDocs and implements the CompileSections postprocessor which will do
the business logic of concatenating multiple OmegaDocs together if those
OmegaDocs describe different sections of the same outputfile.

Additionally, this change adds the GenerateSitemap postrpocessor so that
we can have an entire sitemap generated, linking to all available
documents in all the OmegaDocs.